# src/main/resources/static/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Uniconvert API
  version: v1
  description: |
    Uniconvert backend API documentation (draft).
  contact:
    name: Uniconvert
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
  /auth/signup:
    post:
      tags:
        - Auth
      summary: 자체 회원가입 | 텐텐
      description: |
        [확인 필요]
        - Figma에는 약관 동의가 "약관동의" 화면으로 별도 분리되어 있습니다. 약관 동의 데이터가
          이 API 요청(terms 배열)에 함께 전달되는지, 아니면 별도 API/스텝으로 저장되는지 확정 필요.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - terms
              properties:
                email:
                  type: string
                  format: email
                  description: 로그인 이메일
                password:
                  type: string
                  format: password
                  description: 영문·숫자·특수문자 포함
                terms:
                  type: array
                  description: 약관 동의 목록. SERVICE, PRIVACY는 필수 동의.
                  items:
                    type: object
                    required:
                      - termsType
                      - isAgreed
                    properties:
                      termsType:
                        $ref: '#/components/schemas/TermsType'
                      termsVersion:
                        type: string
                        nullable: true
                      isAgreed:
                        type: boolean
      responses:
        '200':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userId:
                            type: integer
                            format: int64
                            description: 생성된 회원 ID
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          isOnboardingCompleted:
                            type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Auth
      summary: 자체 로그인 | 텐텐
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          isOnboardingCompleted:
                            type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/social/google:
    post:
      tags:
        - Auth
      summary: 소셜 로그인/회원가입 (Google/Apple) | 텐텐
      description: |
        [확인 필요]
        - Apple 로그인도 함께 제공할지 (코드의 AuthProvider에는 APPLE도 존재하나 Figma에서는
          Google 버튼만 확인됨)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
                - idToken
              properties:
                provider:
                  $ref: '#/components/schemas/AuthProvider'
                idToken:
                  type: string
                  description: 소셜 provider가 발급한 ID Token
      responses:
        '200':
          description: 소셜 로그인/회원가입 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userId:
                            type: integer
                            format: int64
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          isOnboardingCompleted:
                            type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: 잘못된 요청 (INVALID_REQUEST/VALIDATION_FAILED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: 인증 필요 또는 토큰 만료 (UNAUTHORIZED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
    Conflict:
      description: 중복된 리소스 (DUPLICATE_RESOURCE)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
  schemas:
    ApiResponse:
      type: object
      description: |
        공통 응답 포맷 (com.uniconvert.backend.global.response.ApiResponse 와 동일 구조).
        data는 성공 시 각 API 응답 데이터로 대체되며, 실패 시 null입니다.
      properties:
        success:
          type: boolean
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: 요청이 성공했습니다.
        data:
          type: object
          nullable: true
...
    TermsType:
      type: string
      description: com.uniconvert.backend.domain.auth.enums.TermsType
      enum:
        - SERVICE
        - PRIVACY
        - MARKETING
    AuthProvider:
      type: string
      description: com.uniconvert.backend.domain.auth.enums.AuthProvider
      enum:
        - GOOGLE
        - APPLE
...